Projeto funcioando - Deletei e Recriei a pasta Migrations porque estava inconsistente com a aplicação
