+++ Desafio1/Desafio1/Data/Persistent/ConsultorioContext.cs
@@ -1,5 +1,9 @@
-﻿using Desafio1.Models;
+﻿using Desafio1.Data.Persistent.EntityConfig;
+using Desafio1.Data.Persistent.DbConfig;
+using Desafio1.Models;
 using Microsoft.EntityFrameworkCore;
+using System;
+using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
 namespace Desafio1.Data.Persistent
 {
@@ -8,23 +12,31 @@ namespace Desafio1.Data.Persistent
         public DbSet<Agendamento> Agendamentos { get; set; }
         public DbSet<Paciente> Pacientes { get; set; }
 
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-            // TODO: .env
-            optionsBuilder
-                .UseNpgsql(@$"
-                    Server=babar.db.elephantsql.com;
-                    Database=rodhrtvh;
-                    User Id=rodhrtvh;
-                    Password=TmApvZ5D_Dw4SwU5lCwJUPy77yEfvbyF;
-                    Port=5432")
-                .UseSnakeCaseNamingConvention();
-        }
+        public ConsultorioContext(DbContextOptions<ConsultorioContext> options) : base(options) {}
+
+        // protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+        // {
+        //     optionsBuilder.UseNpgsql(PostgresContextBuilder.Connection());
+        // }
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
             modelBuilder.ApplyConfiguration(new AgendamentoConfiguration());
             modelBuilder.ApplyConfiguration(new PacienteConfiguration());
+        }
+
+        protected sealed override void ConfigureConventions(ModelConfigurationBuilder configurationBuilder)
+        {
+            configurationBuilder.Properties<DateTime>()
+                .HaveConversion(typeof(DateTimeToDateTimeUtc));
+        }
+
+        public class DateTimeToDateTimeUtc : ValueConverter<DateTime, DateTime>
+        {
+            public DateTimeToDateTimeUtc() : base(c => DateTime.SpecifyKind(c, DateTimeKind.Utc), c => c)
+            {
+
+            }
         }
     }
 }
